requirements.txt
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
ultralytics==8.2.78
dj_database_url== 2.2.0
python-dotenv== 1.0.1
psycopg==3.2.1
psycopg2-binary==2.9.9
gunicorn==23.0.0